feat: Refactor position handling by introducing Floor component and updating related systems

+++ Game.ECS/Components/Movement.cs
@@ -4,5 +4,5 @@ namespace Game.ECS.Components;
 // Movement - Movimento
 // ============================================
 public struct Walkable { public float BaseSpeed; public float CurrentModifier; }
-public struct Facing { public int DirectionX; public int DirectionY; }
+public struct Facing { public sbyte DirectionX; public sbyte DirectionY; }
 public struct Movement { public float Timer; } // Acumulador de movimento por célula
